refactor(lsp): rename var_markdown module to description
Both hover content and completion documentation are forms of variable
description, so name the module after that role.
