+++ crates/css-var-kit/src/commands/lsp.rs
@@ -1,12 +1,12 @@
 mod completion;
 mod definition;
+mod description;
 mod diagnostics;
 mod file_watcher;
 mod hover;
 mod logger;
 mod rename;
 mod uri;
-mod var_markdown;
 
 use std::collections::HashMap;
 use std::error::Error;
