+++ crates/css-var-kit/src/commands/lsp/completion.rs
@@ -12,7 +12,7 @@ use lsp_types::{
 };
 
 use super::Server;
-use super::var_markdown::{
+use super::description::{
     format_multi_def, format_single, multi_def_separator, resolve_to_raw_value,
 };
 use crate::searcher::conditions::variable_definitions::VariableDefinitions;
